armv7m: add armv7m_excp_running_prio()
Add prigroup to CPU environment and ensure that
BASEPRI respects it (via MSR).

Implements v7m exception priority algorithm
using FAULTMASK, PRIMASK, BASEPRI, and the highest
priority active exception.

The number returned is the current execution priority
which may be in the range [-2,0xff] when an exception is active
or 0x100 when no exception is active.