armv7m: remove extra cpu_reset()
cpu_reset() is called as a side-effect of realizing the CPU.
arm_cpu_reset() calls rom_ptr(0), which expects to find
the image mapped.  This was happening way before load_*()
and was worked around with a second call to cpu_reset().
Now wait to realize until after the image is mapped.

Signed-off-by: Michael Davidsaver <mdavidsaver@gmail.com>